requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
requests==2.31.0
Pillow==10.2.0
python-dotenv==1.0.0 